reorganize functions in util and moment files
